// src/app/models/UpdateFOS.ts
export class UpdateFOS {
    api_token: any;
    prefix: any;
    first_name: any;
    middle_name: any;
    last_name: any;
    country_code = '';
    mobile_number: any;
    email: any;
    password:any;
    confirm_password:any;
    cp_executive_id = '';
    is_team_lead: any= 0;
    is_admin: any= 2;
}